# apis/backend-api/web-framework/openapi.json
{
  "components": {
    "schemas": {
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "format": "int32",
            "maximum": 40099,
            "minimum": 40011,
            "type": "integer"
          },
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
      "PageInfo": {
        "additionalProperties": false,
        "properties": {
          "hasMore": {
            "type": "boolean"
          },
          "mode": {
            "enum": [
              "offset",
              "cursor"
            ],
            "type": "string"
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "page": {
            "minimum": 1,
            "type": "integer"
          },
          "pageSize": {
            "maximum": 200,
            "minimum": 1,
            "type": "integer"
          },
          "totalItems": {
            "pattern": "^[0-9]+$",
            "type": "string"
          },
          "totalPages": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "mode"
        ],
        "type": "object"
      },
      "ProblemDetail": {
        "additionalProperties": true,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "instance": {
            "type": "string"
          },
          "status": {
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "traceId": {
            "description": "Server-owned request correlation id.",
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "format": "uri-reference",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "type": "object"
      },
      "SdkWorkApiResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "default": 0,
            "enum": [
              0
            ],
            "format": "int32",
            "maximum": 0,
            "minimum": 0,
            "type": "integer"
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "description": "Server-owned request correlation id.",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "type": "object"
      },
...
      "SdkWorkListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SdkWorkPageData"
              }
            },
            "required": [
              "data"
            ],
            "type": "object"
          }
        ]
      },
      "SdkWorkPageData": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array"
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "items",
          "pageInfo"
        ],
        "type": "object"
      },
      "SdkWorkPlatformErrorCode": {
        "description": "Platform or domain error code per API_SPEC.md §15.3.",
        "format": "int32",
        "maximum": 79999,
        "minimum": 40001,
        "type": "integer"
      },
      "SdkWorkResourceData": {
        "additionalProperties": false,
        "properties": {
          "item": {
            "additionalProperties": true,
            "description": "Typed domain resource for the operation.",
            "type": "object"
          }
        },
        "required": [
          "item"
        ],
        "type": "object"
      },
      "SdkWorkResourceResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SdkWorkResourceData"
              }
            },
            "required": [
              "data"
            ],
            "type": "object"
          }
        ]
      }
    },
    "securitySchemes": {
      "sdkworkAccessToken": {
        "description": "Signed JWT access_token (`header.payload.signature`). Required on all non-open-api routes, including public and refresh-token entrypoints, for tenant isolation.",
        "in": "header",
        "name": "Access-Token",
        "type": "apiKey"
      },
      "sdkworkAgentToken": {
        "description": "Backend agent bootstrap token for /backend/v3/api/agent/* routes. Resolves via api-key auth-mode without dual-token JWT.",
        "in": "header",
        "name": "X-SDKWork-Agent-Token",
        "type": "apiKey"
      },
      "sdkworkDualToken": {
        "description": "Protected routes require Authorization: Bearer <auth_token JWT> and Access-Token: <access_token JWT>.",
        "in": "header",
        "name": "Access-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "SDKWork Web Framework Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/backend/v3/api/web-framework/cors_policies": {
      "get": {
        "operationId": "webFramework.corsPolicies.list",
        "parameters": [
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkListResponse"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "sdkworkDualToken": []
          }
        ],
        "summary": "webFramework.corsPolicies.list",
        "tags": [
          "WebFramework"
        ],
        "x-sdkwork-alternate-permissions": [
          "web-framework.platform.read"
        ],
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "web-framework.tenant.admin",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-required-surface": "organizationMember",
        "x-sdkwork-route-auth": "dual-token"
      },
      "put": {
        "operationId": "webFramework.corsPolicies.upsert",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkResourceResponse"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Forbidden"
          },
          "413": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Payload Too Large"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "sdkworkDualToken": []
          }
        ],
        "summary": "webFramework.corsPolicies.upsert",
        "tags": [
          "WebFramework"
        ],
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "web-framework.tenant.admin",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-required-surface": "organizationMember",
        "x-sdkwork-route-auth": "dual-token"
      }
    },
...
  }
}